[김정연] 강의담당관리 경로 설정, 공통코드 입력, 검색 기능 추가, 수업관리 코드 입력 기능추가.
1. 강의 담당관리 경로 설정.


	+module 내에 sidebar.jsp 화면에 강의담당관리 경로 설정.



2. 공통코드 입력, 검색기능 추가.

	+view페이지 내에 commonCode 폴더에 form 폴더 추가. 그 내부에 공통코드 입력 폼 추가.

	+CommonCodeController, CommonCodeDao, CommonCodeService,
CommonCodeMapper.xml 내용 추가 -> 입력 처리 기능 추가.

	+view 페이지 내 공통코드 리스트 상단에 검색 창 추가 및, 제이쿼리를 이용한 검색기능 추가.



3. 수업관리 코드 입력 기능 추가.

	+view페이지 내에 lectureCode 폴더에 form 폴더 추가. 그 내부에 수업관리 코드 입력 폼 추가.

	+ManageLectureController, ManageLectureCodeDao, ManageLectureCodeMa
pper, ManageLectureCodeService 내용 추가 -> 입력처리 기능 추가.




4. 기타.

	+GradingCriteriaCode 학점이 저장되는 변수의 데이터 타입을 double로 바꿈.

	+view 페이지의 공통코드 리스트 보여줄 테이블 수정

+++ haksaProject/src/main/java/com/cafe24/iumium/manage/lecture/dto/GradingCriteriaCode.java
@@ -5,7 +5,7 @@ public class GradingCriteriaCode {
 	private String recordId;
 	private int gradingCriteriaStartOfSection;
 	private int gradingCriteriaEndOfSection;
-	private int gradingCriteriaGrade;
+	private double gradingCriteriaGrade;
 	public String getGradingCriteriaRating() {
 		return gradingCriteriaRating;
 	}
@@ -30,10 +30,10 @@ public class GradingCriteriaCode {
 	public void setGradingCriteriaEndOfSection(int gradingCriteriaEndOfSection) {
 		this.gradingCriteriaEndOfSection = gradingCriteriaEndOfSection;
 	}
-	public int getGradingCriteriaGrade() {
+	public double getGradingCriteriaGrade() {
 		return gradingCriteriaGrade;
 	}
-	public void setGradingCriteriaGrade(int gradingCriteriaGrade) {
+	public void setGradingCriteriaGrade(double gradingCriteriaGrade) {
 		this.gradingCriteriaGrade = gradingCriteriaGrade;
 	}
 }
